+++ ui/components/ui/entities/get-provider-logo.tsx
@@ -8,6 +8,7 @@ import {
   KS8ProviderBadge,
   M365ProviderBadge,
   MongoDBAtlasProviderBadge,
+  OpenStackProviderBadge,
   OracleCloudProviderBadge,
 } from "@/components/icons/providers-badge";
 import { ProviderType } from "@/types";
@@ -34,6 +35,8 @@ export const getProviderLogo = (provider: ProviderType) => {
       return <MongoDBAtlasProviderBadge width={35} height={35} />;
     case "alibabacloud":
       return <AlibabaCloudProviderBadge width={35} height={35} />;
+    case "openstack":
+      return <OpenStackProviderBadge width={35} height={35} />;
     default:
       return null;
   }
@@ -61,6 +64,8 @@ export const getProviderName = (provider: ProviderType): string => {
       return "MongoDB Atlas";
     case "alibabacloud":
       return "Alibaba Cloud";
+    case "openstack":
+      return "OpenStack";
     default:
       return "Unknown Provider";
   }